requirements.txt
customtkinter
Pillow
threading
textract
transformers
numpy
openai
faiss
python-dotenv
